+++ cpr/session.cpp
@@ -1,4 +1,4 @@
-#include "session.h"
+#include "cpr/session.h"
 
 #include <algorithm>
 #include <functional>
@@ -6,8 +6,8 @@
 
 #include <curl/curl.h>
 
-#include "curlholder.h"
-#include "util.h"
+#include "cpr/curlholder.h"
+#include "cpr/util.h"
 
 namespace cpr {
