+++ include/cpr.h
@@ -1,11 +1,10 @@
 #ifndef CPR_CPR_H
 #define CPR_CPR_H
 
-#include "api.h"
-#include "auth.h"
-#include "cprtypes.h"
-#include "response.h"
-#include "session.h"
-
+#include "cpr/api.h"
+#include "cpr/auth.h"
+#include "cpr/cprtypes.h"
+#include "cpr/response.h"
+#include "cpr/session.h"
 
 #endif
